feat(license): Move Dependency tests to new file
